% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addSentimentToDf.R
\name{addSentimentToDf}
\alias{addSentimentToDf}
\title{A function that adds sentiment columns to df}
\usage{
addSentimentToDf(df, column = NULL)
}
\arguments{
\item{df}{a dataframe}

\item{column}{a text column of dataframe df}
}
\value{
a dataframe with calculated sentiment is returned- the positive, negative and overall score
}
\description{
The function below calculates the sentiment scores of texts and returns a dataframe with
collected sentiments added to existing dataframe
}
\examples{
addSentimentToDf(data.frame(Review.Text = c("I love strawberries", "I hate chocolate")))
}
\author{
Zuzanna Miazio Karolina Solarska
}
