% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAllWOrdsSentiments.R
\name{getAllWOrdsSentiments}
\alias{getAllWOrdsSentiments}
\title{A function that adds sentiment columns to df}
\usage{
getAllWOrdsSentiments(column)
}
\arguments{
\item{column}{a text column of dataframe df}
}
\value{
a dataframe with calculated sentiment is returned- the positive, negative
}
\description{
The function below calculates the sentiment scores of texts and returns a dataframe with
collected sentiments added to existing dataframe
}
\examples{
getAllWOrdsSentiments(c("I love strawberries", "I hate chocolate"))
}
\author{
Zuzanna Miazio Karolina Solarska
}
