% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetNegativeSentiment.R
\name{GetSentimentNeg}
\alias{GetSentimentNeg}
\alias{GetNegativeSentiment}
\title{A function that calculates sentiment based on a dataframe column}
\usage{
GetNegativeSentiment(file)
}
\arguments{
\item{file}{text column of a dataframe}
}
\value{
a dataframe with calculated sentiment is returned- the positive, negative and overall score
}
\description{
The function below calculates the sentiment scores of texts and returns a dataframw eith collected sentiments
}
\examples{
GetPositiveSentiment("I like ice cream, I hate chocolate")
}
\author{
Zuzanna Miazio Karolina Solarska
}
