\docType{package}
\name{sentiment-package}
\alias{sentiment-package}
\title{Calculates sentiment scores using affinity dictionaries.}
\description{
  Calculates sentiment scores using affinity dictionaries.
}
\details{
  This is a set of tools for calculating sentiment scores
  using affinity dictionaries.

  Export a function \code{\link{sentiment}}.
}
\author{
  Andrie de Vries \email{apdevries@gmail.com}
}
\seealso{
  \code{\link{sentiment}}
}
\keyword{package}

